Fix wrong test for submission status
The old test was always true (3 !== STATUS_PUBLISHED is always true),
so the conditional block was always entered.

Then a second bug was triggered if there was no user. Fix that, too.

Now it is possible to see or download citations for users who are
not registered or who do not have any special right.

Signed-off-by: Stefan Weil <sw@weilnetz.de>