Update ABNT-NBR 6023.2002 to 2018
Updating to the latest ABNT standard

According to a librarian specialized in this standard, https://www.zotero.org/styles/universidade-de-sao-paulo-escola-de-comunicacoes-e-artes-abnt is a more correct and current implementation. Change only in the title to make it clear that it is the updated standard and not the name of the institution that updated this template.